Bill S5089 version original - Cordell Cleare
BILL NUMBER: S5089 SPONSOR: CLEARE TITLE OF BILL: An act to amend the public officers law, in relation to the unauthorized release of sealed records PURPOSE: To establish the crime of intentionally contents of a sealed record. SUMMARY OF PROVISIONS: Section one of the bill adds 6 72-b to the public officers law relating to the unauthorized release of sealed records.