Bill S5038 version original - Andrew J. Lanza
BILL NUMBER: S5038 SPONSOR: LANZA TITLE OF BILL: An act to amend the public officers law, in relation to certain appoint- ments by the governor and the senate PURPOSE OF THE BILL: This bill would amend the public officers law to ensure the legislature is given an appropriate amount of time to confirm and reject such appointments, and given an appropriate amount of time to hold a hearing on such appointments. SUMMARY OF SPECIFIC PROVISIONS: Section 1.