Bill S5098 version original - Cordell Cleare
BILL NUMBER: S5098 SPONSOR: CLEARE TITLE OF BILL: An act to amend the criminal procedure law and the correction law, in relation to notifying individuals in pre-trial detention of their right to vote PURPOSE: To ensure that people who are in pre-trial detention and who have not been convicted are made aware of their right to vote. SUMMARY OF PROVISIONS: Section 1 adds new subdivision 4 to S 210.