Bill S5031 version original - Jabari Brisport
BILL NUMBER: S5031 SPONSOR: BRISPORT TITLE OF BILL: An act to amend the social services law, in relation to limitations on the number of cases handled by child welfare personnel SUMMARY OF PROVISIONS: Section 1 amends section 409-g of the social services law to limit the number of cases on a foster care worker's caseload to a maximum of 16. Section 2 provides for an effective date of 60 days after the bill becomes law.