Bill S5026 version original - Andrew Gounardes
BILL NUMBER: S5026 SPONSOR: GOUNARDES TITLE OF BILL: An act to amend the labor law, in relation to enacting the "freelance isn't free act" PURPOSE OR GENERAL IDEA OF BILL: To provide recourse to freelance workers experiencing wage theft and other violations under the Labor Law SUMMARY OF PROVISIONS: Section one of the bill names it the Freelance Isn't Free Act. Section two of the bill adds a new section 191-d to the Labor Law.