Bill S5030 version original - Jabari Brisport
BILL NUMBER: S5030 SPONSOR: BRISPORT TITLE OF BILL: An act to amend the agriculture and markets law, in relation to requir- ing the owner of any place or establishment where animals or fowls are slaughtered or butchered for food to reimburse costs associated with capture and return or rescue and rehabilitation of escaped animals SUMMARY OF PROVISIONS: Section one amends the agriculture and markets law by adding a new section 96-e-1. Section two sets forth the effective date.