Bill S5075 version original - Cordell Cleare
BILL NUMBER: S5075 SPONSOR: CLEARE TITLE OF BILL: An act to amend the real property tax law and the administrative code of the city of New York, in relation to prohibiting the sale of tax liens by a tax district in a city with a population of one million or more PURPOSE OR GENERAL IDEA OF BILL: Prohibits tax districts from selling their tax liens to private collec- tors, requiring municipalities to collect their own debts. SUMMARY OF SPECIFIC PROVISIONS: Section 1.