Bill S5033 version original - Jabari Brisport
BILL NUMBER: S5033 SPONSOR: BRISPORT TITLE OF BILL: An act to amend the civil practice law and rules, in relation to the commencement of an action on a money judgment; and to repeal certain provisions of such law relating thereto SUMMARY OF PROVISIONS: Section one of this bill repeals subdivision (b) of section 211 of the civil practice law and rules. Section two provides the timeframe for collecting a money judgment under New York law.