Bill S5064 version original - Cordell Cleare
BILL NUMBER: S5064 SPONSOR: CLEARE TITLE OF BILL: An act to amend the executive law, in relation to prohibiting the invol- untary use of sedatives during an arrest PURPOSE OR GENERAL IDEA OF BILL: To stop police from tranquilizing people so that they can be arrested and to avoid unnecessary deaths. SUMMARY OF SPECIFIC PROVISIONS: Section 1.