Bill S5107C version original - Brad Hoylman-Sigal
BILL NUMBER: S5107 SPONSOR: HOYLMAN-SIGAL TITLE OF BILL: An act to amend the family court act, the general business law and the domestic relations law, in relation to surrogacy programs and agreements PURPOSE: To make necessary technical corrections to Part L of chapter 56 of the laws of 2020. SUMMARY OF PROVISIONS: Sections 1 through 21 make technical corrections to Article 5-C of the Family Court Act. Section 22 makes a technical correction to section 123(2)(a) of the domestic relations law.