Bill S5128 version original - Lea Webb
BILL NUMBER: S5128 SPONSOR: WEBB TITLE OF BILL: An act to amend the vehicle and traffic law, in relation to establishing a distinctive plate honoring recipients of the Girl Scout Gold Award PURPOSE: Amends the Vehicle and Traffic Law to allow for distinctive plates for recipients of the Girl Scout Gold Award. SUMMARY OF PROVISIONS: Section 1.