Bill S5151 version original - Kevin S. Parker
BILL NUMBER: S5151 SPONSOR: PARKER TITLE OF BILL: An act to amend the executive law, in relation to the siting of the New York state Martin Luther King, Jr. institute for nonviolence at Medgar Evers college PURPOSE: The purpose of this bill is to permit the siting of the Martin Luther King, Jr. Institute for Nonviolence at Medgar Evers College in Brooklyn, New York.