Bill S5112 version original - Peter Oberacker
BILL NUMBER: S5112 SPONSOR: OBERACKER TITLE OF BILL: An act to amend the education law, in relation to creating a school resource officer working group; and providing for the repeal of such provisions upon the expiration thereof SUMMARY OF PROVISIONS: Section 1. Adds Section 2801-c to the Education Law providing for the creation of a working group to examine the need for standards and train- ing for school resource officers.