Bill S5146 version original - Kevin S. Parker
BILL NUMBER: S5146 SPONSOR: PARKER TITLE OF BILL: An act to amend the mental hygiene law, in relation to establishing a peer support program for veterans PURPOSE OR GENERAL IDEA OF BILL: The purpose of this bill is to establish e peer support program for veterans. Develop and establish regional and county based peer support services programs for women veterans and establishes minimum qualifica- tions for recognized or certified veteran peer counselor.