Bill S5119A version original - Pamela Helming
BILL NUMBER: S5119 SPONSOR: HELMING TITLE OF BILL: An act to amend the tax law, in relation to authorizing the county of Livingston to impose an additional one percent sales tax PURPOSE: Relates to extending the authorization for Livingston County to impose additional rates of sales and compensating use taxes. SUMMARY OF PROVISIONS: Section 1.