Bill S5140 version original - Kevin S. Parker
BILL NUMBER: S5140 SPONSOR: PARKER TITLE OF BILL: An act to amend the labor law, in relation to clarifying the definition of a "commission salesperson" PURPOSE: To clarify the definition of who is a "Commission salesman" under the New York labor law. SUMMARY OF PROVISIONS: Section 1: Subdivision 6 of section 190 of the labor law, as added by chapter 548 of the laws 1966, is amended. Section 2: This act shall take effect immediately.