Bill S238A version A - Rachel May
BILL NUMBER: S238A SPONSOR: MAY TITLE OF BILL: An act to amend the environmental conservation law and the state finance law, in relation to establishing the public water justice act PURPOSE: To recognize the public's interest in protecting New York's freshwater resources and establish fees for the bottling and sale of the state's water by commercial entities. SUMMARY OF PROVISIONS: Section 1: Sets the name of the legislation as the Public Water Justice Act.