Bill S5110 version original - Peter Oberacker
BILL NUMBER: S5110 SPONSOR: OBERACKER TITLE OF BILL: An act to amend the real property tax law, in relation to extending the benefits of the STAR program to small businesses PURPOSE: Allow real property owned by a small business to be eligible for the STAR program. SUMMARY OF PROVISIONS: Section 1 Amends the real property tax law to include small businesses to qualify for the STAR property tax exemption. Defines small business to mean a business which employs one hundred persons or less.