Bill S5123 version original - Andrew J. Lanza
BILL NUMBER: S5123 SPONSOR: LANZA TITLE OF BILL: An act to amend the general municipal law and the administrative code of the city of New York, in relation to the regulation of fifth generation (5G) telecommunications towers in cities with a population of one million or more PURPOSE: To make the siting of 5G towers in New York City transparent to the public and environmentally safe to communities. SUMMARY OF SPECIFIC PROVISIONS: Section 1 adds a new section 137 to the general municipal law.