Bill S5137 version original - Kristen Gonzalez
BILL NUMBER: S5137 SPONSOR: GONZALEZ TITLE OF BILL: An act to amend the civil practice law and rules, in relation to certif- ication of class actions in cases involving governmental operations PURPOSE: The civil practice law and rules is amended to provide that a court may not deny or withhold class certification solely because the action involves governmental operations. This bill will not require the court to certify a class.