Bill S5149 version original - Kevin S. Parker
BILL NUMBER: S5149 SPONSOR: PARKER TITLE OF BILL: An act to amend the education law, in relation to documenting campus sexual assault and related complaints at SUNY and CUNY institutions PURPOSE: To address the alarming increase of sexual assaults on public university and college campuses. SUMMARY OF SPECIFIC PROVISIONS: Section 1: Section 6433 of the education law, as added by chapter597 of the laws of 2003, is amended.