Bill S5133 version original - Roxanne J. Persaud
BILL NUMBER: S5133 SPONSOR: PERSAUD TITLE OF BILL: An act to amend the tax law and the state finance law, in relation to providing for taxpayer gifts for uterine fibroids education and estab- lishing the uterine fibroids education fund PURPOSE: Establishes non-exempt taxpayer-funded gifts toward the uterine fibroids education fund. Requires the commissioner to include space on the personal income tax return to enable a taxpayer to contribute to the fund.