Bill S5155 version original - Kevin S. Parker
BILL NUMBER: S5155 SPONSOR: PARKER TITLE OF BILL: An act to amend the judiciary law, in relation to enacting the "language barrier to justice elimination act" PURPOSE: To establish guidelines for court interpreters. SUMMARY OF SPECIFIC PROVISIONS: Section 1: This act shall be known and may be cited as the "language barrier to justice elimination act". Section 2: Section 386 of the judiciary law, as amended by chapter 649 of the laws of 1945, is amended.