Bill S5145 version original - Kevin S. Parker
BILL NUMBER: S5145 SPONSOR: PARKER TITLE OF BILL: An act to amend the banking law, in relation to credit union memberships and general powers PURPOSE OR GENERAL IDEA OF BILL: This bill would expand the qualifications for membership into credit unions. SUMMARY OF SPECIFIC PROVISIONS: Section 1: Subdivision 2 of section 451 of the banking law, as amended by chapter 502 of the laws of 2014, is amended.