Bill S5139 version original - Kevin S. Parker
BILL NUMBER: S5139 SPONSOR: PARKER TITLE OF BILL: An act to amend the legislative law, in relation to creating a cost of living adjustment to be applied to the salaries of members of the legis- lature PURPOSE: To adjust the salaries of members of the legislature in congruence with the cost of living adjustment. SUMMARY OF PROVISIONS: Section 1: Section 5-a of the legislative law is amended by adding a new subdivision 4. Section 2: This act shall take effect immediately.