Bill S5183 version original - Robert Ortt
BILL NUMBER: S5183 SPONSOR: ORTT TITLE OF BILL: An act to amend the tax law, in relation to extending the period during which the county of Orleans is authorized to impose additional rates of sales and compensating use taxes PURPOSE: This legislation would extend the sales and compensating use tax in the county of Orleans. SUMMARY OF PROVISIONS: Section 1.