Bill S5223 version original - Kevin S. Parker
BILL NUMBER: S5223 SPONSOR: PARKER TITLE OF BILL: An act to amend the executive law, in relation to requiring state and local law enforcement officers to identify themselves to the public PURPOSE OR GENERAL IDEA OF BILL: To increase transparency in police practices and rebuild trust between police officers and members of the public. SUMMARY OF SPECIFIC PROVISIONS: Section 1: Declaration of legislative intent and findings. Section 2: The executive law is amended by adding a new article 13-B.