Bill S5215 version original - Dean Murray
BILL NUMBER: S5215 SPONSOR: MURRAY TITLE OF BILL: An act authorizing the town of Brookhaven, county of Suffolk to amend a conservation easement PURPOSE: To authorize the town of Brookhaven, county of Suffolk to adjust the boundaries of a conservation easement for the express purpose of allow- ing rail access and expanding the lands of a conservation easement. SUMMARY OF PROVISIONS: Section 1.