Bill S5198 version original - James Skoufis
BILL NUMBER: S5198 SPONSOR: SKOUFIS TITLE OF BILL: An act to amend the real property law, in relation to increases of rent in manufactured home parks PURPOSE: Clarify existing requirements of manufactured home park owners to justi- fy rent increases in excess of 3% of the current rent. SUMMARY OF PROVISIONS: Section 1: Amends § 233-b of the Real Property Law to add a new subdivi- sion 9 and subdivision 10.