Bill S5177A version A - Dean Murray
BILL NUMBER: S5177A SPONSOR: MURRAY TITLE OF BILL: An act to amend the public health law and the election law, in relation to notifying county boards of elections of an individual's death PURPOSE: To ensure that upon the death of a registered voter of New York state, that the county boards of elections are notified in a timely manner so that the voter registration rolls may be more accurately maintained. SUMMARY OF PROVISIONS: Section 1.