Bill S5210 version original - Patrick M. Gallivan
BILL NUMBER: S5210 SPONSOR: GALLIVAN TITLE OF BILL: An act to amend the tax law, in relation to exempting railroad rolling stock from sales and compensating use taxes PURPOSE: To amend the tax law, creating a sales and use tax exemption for manufacturers of railroad rolling stock. SUMMARY OF PROVISIONS: Section 1. Provides for the amending of the tax law to create the sales and use tax exemption and to define "railroad rolling stock" for the purpose of this section.