Bill S5202 version original - James Skoufis
BILL NUMBER: S5202 SPONSOR: SKOUFIS TITLE OF BILL: An act to amend the general business law, in relation to fines for false alarms by an alarm system PURPOSE: To permit the levying of fines against alarm system companies solely for false alarms attributed to attributed to a deficiency in the alarm system or an error of the alarm system company or central station. SUMMARY OF PROVISIONS: Section 1: amends section 69-1 of the general business law to add new subdivisions 6, 7, 8, and 9.