Bill S4168B version A - Anthony H. Palumbo
BILL NUMBER: S4168A SPONSOR: PALUMBO TITLE OF BILL: An act to amend chapter 672 of the laws of 1993, amending the public authorities law relating to the construction and financing of facilities for certain public libraries, in relation to including the Shelter Island Public Library PURPOSE: This bill would add the Shelter Island Public Library Society to the list of entities eligible for DASNY services. SUMMARY OF PROVISIONS: Section 1.