Bill S5226 version original - Kevin S. Parker
BILL NUMBER: S5226 SPONSOR: PARKER TITLE OF BILL: An act to amend the tax law, in relation to the creation of the empire state music production credit PURPOSE OR GENERAL IDEA OF BILL: Relates to the creation of the empire state music production tax credit. SUMMARY OF SPECIFIC PROVISIONS: The tax law is amended by adding a new section 42. Section 210-B of the tax law is amended by adding a new subdivision 49.