Bill S5200 version original - James Skoufis
BILL NUMBER: S5200 SPONSOR: SKOUFIS TITLE OF BILL: An act to amend the public health law, in relation to adding substances to the schedules of controlled substances PURPOSE: Adds certain substances to the schedules of controlled substances. SUMMARY OF PROVISIONS: Section 1: Amends subdivision (b) of schedule I of section 3306 of the public health law by adding four new paragraphs 71, 72,73 and 74.