Bill S5225 version original - Kevin S. Parker
BILL NUMBER: S5225 SPONSOR: PARKER TITLE OF BILL: An act to amend the education law, in relation to screening of students for eating disorders PURPOSE OR GENERAL IDEA OF BILL: To require public school students to be screened for eating disorders. SUMMARY OF SPECIFIC PROVISIONS: Section 1: Subdivision 1 of section 903 of the education law, as amended by chapter 183 of the laws of 2017, is amended.