Bill S5247 version original - Alexis Weik
BILL NUMBER: S5247 SPONSOR: WEIK TITLE OF BILL: An act to amend the criminal procedure law, in relation to protecting the identity of witnesses of hate crimes in discovery proceedings SUMMARY OF PROVISIONS: Section 1. Amends paragraphs (c) and (g) of subdivision I of section 245.20 of the criminal procedure law, as amended by section 2 of part HHH of chapter 56 of the laws of 2020, to protect the identity of a witness of a hate crime, as defined under section 485.