Bill S5256 version original - James Sanders Jr.
BILL NUMBER: S5256 SPONSOR: SANDERS TITLE OF BILL: An act to amend the general business law, in relation to prohibiting the inclusion of a confession of judgment in a contract or agreement for a financial product or service PURPOSE OR GENERAL IDEA OF BILL: To prohibit "confessions of judgment" in any contract or agreement for a financial product or service. SUMMARY OF PROVISIONS: Section 1 - Prohibits "confession of judgment" for any contract or agreement for 'a financial product or service.