Bill S5254 version original - Robert Jackson
BILL NUMBER: S5254 SPONSOR: JACKSON TITLE OF BILL: An act to amend the retirement and social security law, in relation to providing accidental disability retirement for deputy sheriffs in Nassau county PURPOSE: To provide the option for an accidental disability retirement benefit of 75% of the final average salary for deputy sheriffs employed by Nassau County. SUMMARY OF PROVISIONS: Section 1 adds a new Section 605-g to the Retirement and Social Security Law.