Bill S5251 version original - Alexis Weik
BILL NUMBER: S5251 SPONSOR: WEIK TITLE OF BILL: An act to amend the penal law, in relation to establishing the crime of assault on a pregnant individual PURPOSE: To establish the crime of assault on a pregnant individual. SUMMARY OF PROVISIONS: Section one states the title of the Liv Act. Section two adds new section 120.09 to the penal code. It provides that a person who knowingly or intentionally causes injury to pregnant indi- vidual shall be guilty of assault.