Bill S2163B version A - Joseph P. Addabbo Jr.
BILL NUMBER: S2163A SPONSOR: ADDABBO TITLE OF BILL: An act to amend the agriculture and markets law, in relation to prohib- iting the slaughtering of horses for human or animal consumption PURPOSE OR GENERAL IDEA OF BILL: To amend the Agriculture and Markets Law by adding a new section 385, to prohibit the slaughtering of horses for human or animal consumption. SUMMARY OF SPECIFIC PROVISIONS: Section 1: The Agriculture and Markets Law shall be amended by adding a new section 385.