Bill S5315 version original - Alexis Weik
BILL NUMBER: S5315 SPONSOR: WEIK TITLE OF BILL: An act to amend the executive law, in relation to requiring unanimous agreement by the parole board to release an incarcerated individuals on parole PURPOSE: To require unanimous consent of the Parole Board to release an inmate on parole. SUMMARY OF PROVISIONS: Sections one and two of this bill amend Executive Law section 259(2)(a) to require that decisions of the Parole Board to grant parole to an inmate must be unanimous.