Bill S5317 version original - Alexis Weik
BILL NUMBER: S5317 SPONSOR: WEIK TITLE OF BILL: An act to amend the highway law, in relation to designating a portion of the state highway system as the "Sergeant Gregory Wahl Memorial Highway" PURPOSE OR GENERAL IDEA OF BILL: To designate a portion of the state highway system as the "Sergeant Gregory Wahl Memorial". SUMMARY OF SPECIFIC PROVISIONS: Section 1. Amends the highway law by designating a portion of the state highway system as "Sergeant Gregory Wahl Memorial Highway." Section 2.