Bill S5282 version original - Kevin S. Parker
BILL NUMBER: S5282 SPONSOR: PARKER TITLE OF BILL: An act to amend the domestic relations law, in relation to establishing a presumption of shared parenting of minor children in matrimonial proceedings PURPOSE: To create a statutory presumption of joint custody for all minor chil- dren whose parents are no longer married, so that both parents can continue to share in the responsibilities and duties of the children's. upbringing. SUMMARY OF PROVISIONS: Section 1: Legislative findings.