Bill S5267 version original - Cordell Cleare
BILL NUMBER: S5267 SPONSOR: CLEARE TITLE OF BILL: An act to amend the correction law, in relation to providing for the creation and maintenance of digital records for each incarcerated indi- vidual PURPOSE OR GENERAL IDEA OF BILL: Provides for the creation and maintenance of digital records for each incarcerated individual. SUMMARY OF PROVISIONS: Section 1 Section 112 of the correction law is amended by adding a new subdivision 6.