Bill S5305 version original - Nathalia Fernandez
BILL NUMBER: S5305 SPONSOR: FERNANDEZ TITLE OF BILL: An act to amend the parks, recreation and historic preservation law, in relation to requiring solar panels on installed light fixtures PURPOSE OR GENERAL IDEA OF BILL: This legislation will require solar panels to be installed on light fixtures in state-run parks and parkways. SUMMARY OF PROVISIONS: Section 1: The parks, recreation and historic preservation law is amended by adding a new section 13.