Bill S5292 version original - Nathalia Fernandez
BILL NUMBER: S5292 SPONSOR: FERNANDEZ TITLE OF BILL: An act to amend the public health law, in relation to the definition of elevated blood lead levels PURPOSE OR GENERAL IDEA OF BILL: This bill will lower the numeric definition of elevated lead levels in public health law from 5 to 3.5 micrograms. SUMMARY OF PROVISIONS: Section 1 amends the public health law to change the definition of "elevated lead levels." Section 2 requires the DOH to adopt all necessary regulations related to this change.