Bill S5314 version original - Alexis Weik
BILL NUMBER: S5314 SPONSOR: WEIK TITLE OF BILL: An act to amend the criminal procedure law, in relation to protecting vulnerable witness and victim's contact information PURPOSE: To ensure privacy for vulnerable witnesses and victims of crimes. SUMMARY OF PROVISIONS: Section 1: Provides that the disclosure of names and contact information for persons the court finds on the record are subject to potential witness intimidation, violence or threats is not required.