Bill S5276 version original - Kevin S. Parker
BILL NUMBER: S5276 SPONSOR: PARKER TITLE OF BILL: An act to amend the tax law, in relation to creating a disabled person retrofit tax credit PURPOSE: To provide an incentive for owners of residential buildings to retrofit homes for access to the disabled. SUMMARY OF PROVISIONS: Section 1: Section 606 of the tax law is amended by adding a new section (000).