Bill S5309 version original - Alexis Weik
BILL NUMBER: S5309 SPONSOR: WEIK TITLE OF BILL: An act to amend the family court act, the criminal procedure law and the domestic relations law, in relation to an order of protection with respect to companion animals PURPOSE: To extend protections to the pets of victims of domestic abuse by giving the court discretion to forbid contact between the abuser and any pet cared for by the victim. SUMMARY OF PROVISIONS: Section 1: Amends section 352.