Bill S5270A version original - Roxanne J. Persaud
BILL NUMBER: S5270 SPONSOR: PERSAUD TITLE OF BILL: An act to amend the social services law, in relation to increasing the standards of monthly need and home energy grant amounts for persons in receipt of public assistance PURPOSE: This legislation will increase cash assistance grant amounts for utili- ties and food and other necessities to account for inflation. SUMMARY OF SPECIFIC PROVISIONS: Section one amends paragraph (a-4) of subdivision 2 of section 131-a of the social services law.