Bill S5307 version original - Alexis Weik
BILL NUMBER: S5307 SPONSOR: WEIK TITLE OF BILL: An act to amend the penal law, in relation to aggravated assault in the third degree PURPOSE: This legislation establishes a new crime, aggravated assault in the third degree, when an individual commits the crime of assault in the third degree, as defined under penal law section 120.00, and is aided by another person who is physically present. SUMMARY OF PROVISIONS: The Penal Law is amended by adding a new section 120.