Bill S5316 version original - Alexis Weik
BILL NUMBER: S5316 SPONSOR: WEIK TITLE OF BILL: An act to amend the executive law, in relation to appeals to the state board of parole SUMMARY OF PROVISIONS: Section 1: Subdivision 4 of section 259-1 of the executive law is amended to add a new paragraph (c). Paragraph (c) stipulates that where the board has made a determination granting discretionary release to an individual, the victim of that individual's crime, or their represen- tative, has the right to file an appeal.