Bill S5324 version original - Alexis Weik
BILL NUMBER: S5324 SPONSOR: WEIK TITLE OF BILL: An act to amend the legislative law, in relation to imposing term limits for members of the assembly and senate of the state of New York PURPOSE: Imposes term limits of six two-year terms for members of the Assembly and Senate of New York State. SUMMARY OF PROVISIONS: Section 1: Adds a new section 2-a. Section 2: Establishes the effective date.