Bill S4063A version A - John Mannion
BILL NUMBER: S4063A SPONSOR: MANNION TITLE OF BILL: An act to amend the general business law, in relation to prohibiting publishers from limiting the number of copies of e-books available at public libraries PURPOSE: The purpose of this bill is to prohibit the artificial limitation on the number of copies of e-books available at public libraries. SUMMARY OF PROVISIONS: Section 1 of this bill would add a new section 338 to the General Busi- ness Law.