Bill S5278 version original - Kevin S. Parker
BILL NUMBER: S5278 SPONSOR: PARKER TITLE OF BILL: An act to amend the executive law, in relation to establishing a fatali- ty review board and providing for such board's powers and duties PURPOSE OR GENERAL IDEA OF BILL: To establish an independent fatality review-board consisting of experts to review and assess domestic violence cases which result in a fatality. SUMMARY OF SPECIFIC PROVISIONS: Section 1: Legislative intent. Section 2: The executive law is amended by adding a new section 837-x.