Bill S775A version A - Jeremy Cooney
BILL NUMBER: S775A SPONSOR: COONEY TITLE OF BILL: An act to amend the vehicle and traffic law and the penal law, in relation to the ignition interlock program PURPOSE OR GENERAL IDEA OF BILL: To establish standards and incentives for increased utilization of the ignition interlock device (IID) to protect the public and reduce reci- divism in a manner consistent with best practices. SUMMARY OF PROVISIONS: Section 1: Sets forth the Legislative Findings. Section 2.