Bill S5320 version original - Alexis Weik
BILL NUMBER: S5320 SPONSOR: WEIK TITLE OF BILL: An act to amend the vehicle and traffic law, in relation to renewal of drivers' licenses PURPOSE: To require an updated signature and photo image for renewal and display on a renewed license. SUMMARY OF PROVISIONS: Section 1 amends the vehicle and traffic law to require an updated signature and photo image accompany an application for license renewal. These updates will be displayed on the applicant's new license.