Bill S5341 version original - Joseph P. Addabbo Jr.
BILL NUMBER: S5341 SPONSOR: ADDABBO TITLE OF BILL: An act to amend the general business law, in relation to the licensing of pet grooming facilities PURPOSE OR GENERAL IDEA OF BILL: To require licensing of pet grooming facilities. SUMMARY OF SPECIFIC PROVISIONS: Section one amends the general business law by adding a new article 29-CCC which provides for the licensing of pet grooming facilities. Section 539 of article 29-CCC defines the terms to be used in the arti- cle.