Bill S5330A version original - Pete Harckham
BILL NUMBER: S5330 SPONSOR: HARCKHAM TITLE OF BILL: An act to amend the environmental conservation law, in relation to prohibiting idling by heavy duty vehicles PURPOSE OR GENERAL IDEA OF BILL: To prohibit the idling of on-road heavy duty vehicles, including a bus or truck, for more than 5 consecutive minutes. This bill is designed to improve air quality in New York State.