Bill S4740A version A - James Skoufis
BILL NUMBER: S4740A SPONSOR: SKOUFIS TITLE OF BILL: An act to amend the penal law, in relation to the crime of public lewd- ness PURPOSE: To add cyberflashing to the crime of public lewdness. SUMMARY OF PROVISIONS: Section 1 amends section 245.00 of the penal law by adding a new subdi- vision 2 which adds the act of electronically transmitting an intimate image in or around a public place without consent to the crime of public lewdness. Section 2 provides the effective date.