Bill S5365 version original - Jeremy Cooney
BILL NUMBER: S5365 SPONSOR: COONEY TITLE OF BILL: An act to repeal section 490 of the tax law relating to the excise tax on medical cannabis; and to repeal section 89-h of the state finance law relating to the medical cannabis trust fund SUMMARY OF PROVISIONS: Section 1 repeals section 490 of the tax law. Section 2 repeals section 89-h of the state finance law. Section 3 is the effective date.