Bill S5356 version original - Monica Martinez
BILL NUMBER: S5356 SPONSOR: MARTINEZ TITLE OF BILL: An act to amend the vehicle and traffic law, in relation to road salt PURPOSE: The purpose of this bill is to Permit the transport of road salt as "one piece" or "one item", in order to reduce the number of trucks hauling road salt across our roadways and bridges. SUMMARY OF PROVISIONS: Sections 1. and 2.