Bill S5337 version original - Alexis Weik
BILL NUMBER: S5337 SPONSOR: WEIK TITLE OF BILL: An act to amend the penal law, in relation to adding animal fighting as a criminal act when referring to enterprise corruption PURPOSE: To define animal fighting as a criminal act when referring to enterprise corruption. SUMMARY OF PROVISIONS: Section 1. Amends subdivision 1 of section 460.10 of the penal law to include subdivision 2 of section 351 of the agriculture and markets law as a criminal act when referring to enterprise corruption.