Bill S5340 version original - Alexis Weik
BILL NUMBER: S5340 SPONSOR: WEIK TITLE OF BILL: An act to amend the penal law and the criminal procedure law, in relation to creating the crime of unlawful dissemination or publication of discovery material PURPOSE: To prevent the unlawful dissemination or publication of discovery mate- rial. SUMMARY OF PROVISIONS: Section 1: Amends the penal law by establishing the crime of unlawful dissemination or publication of discovery material as a class E felony.