Bill S2075A version A - Joseph P. Addabbo Jr.
BILL NUMBER: S2075A SPONSOR: ADDABBO TITLE OF BILL: An act to amend the education law, in relation to certain school holi- days; and to repeal certain provisions of such law relating thereto PURPOSE OR GENERAL IDEA OF BILL: To allow the designation of Diwali as a school holiday in New York city. SUMMARY OF PROVISIONS: Section 1 repeals Section 2586 of the education law. Section 2 amends Subdivision 2-a of section 3635 of the education law, as amended by chapter 424 of the laws of 2005.