Bill S2861A version A - Monica Martinez
BILL NUMBER: S2861A SPONSOR: MARTINEZ TITLE OF BILL: An act to amend the family court act, in relation to temporary orders of protection PURPOSE: The purpose of this bill is to allow for an application for a temporary order of protection by the county probation department in the adjustment process. SUMMARY OF PROVISIONS: Section 1. Section 304.