Bill S5345 version original - Michelle Hinchey
BILL NUMBER: S5345 SPONSOR: HINCHEY TITLE OF BILL: An act to amend the parks, recreation and historic preservation law, in relation to the historic business preservation registry PURPOSE: The purpose of this bill is to increase the number of qualified busi- nesses - those in operation for at least fifty years - which an elected official may nominate to the historic business registry each term. SUMMARY OF PROVISIONS: Section 1: Amends subdivision 3 of section 14.