Bill S5357 version original - Lea Webb
BILL NUMBER: S5357 SPONSOR: WEBB TITLE OF BILL: An act to amend the alcoholic beverage control law, in relation to allowing bed and breakfasts to sell cider, liquor, beer and wine PURPOSE: This legislation seeks to expand opportunities for the ability of certain bed and breakfasts to sell alcohol. SUMMARY OF PROVISIONS: Section one: amends subdivision 4 of section 51 of the alcoholic bever- age control law.