Bill S5336 version original - Alexis Weik
BILL NUMBER: S5336 SPONSOR: WEIK TITLE OF BILL: An act to amend the penal law, in relation to vehicular assault and vehicular manslaughter PURPOSE: To amend the penal law, in relation to vehicular assault and vehicular manslaughter SUMMARY OF SPECIFIC PROVISIONS: Section 1. Amends subdivision 3 of section 120.03 of the penal law as amended by Chapter 732 of the Laws of 2006 and adds a new subdivision 4. Section 2. Amends subdivision 5 and 6 of section 125.