Bill S5325 version original - Monica Martinez
BILL NUMBER: S5325 Revised 10/12/23 SPONSOR: MARTINEZ TITLE OF BILL: An act to amend the agriculture and markets law, in relation to sentenc- ing for the crime of aggravated cruelty to animals PURPOSE: To allow for a judge to exercise their discretion to order consecutive or concurrent sentences when an individual is convicted of multiple counts of aggravated cruelty to animals. SUMMARY OF PROVISIONS: Section 1. Names this law "Tucker's Law." Section 2.