Bill S5434 version original - Mario Mattera
BILL NUMBER: S5434 SPONSOR: MATTERA TITLE OF BILL: An act to amend the environmental conservation law, in relation to prohibiting the use of unmanned aerial vehicles in the course of hunting PURPOSE OR GENERAL IDEA OF BILL: To prohibit the use of unmanned aerial vehicles for hunting or taking of wildlife. SUMMARY OF SPECIFIC PROVISIONS: Section 1.