Bill S5381 version original - Pete Harckham
BILL NUMBER: S5381 SPONSOR: HARCKHAM TITLE OF BILL: An act to amend the education law, in relation to the institution of a screening, brief intervention and referral to treatment/evaluation (SBIRT/SBIRE) at colleges PURPOSE: To ensure higher education institutions provide screening, brief intro- duction, and referral to treatment/evaluation(SBIRT/SBIRE) on campus. SUMMARY OF PROVISIONS: Section 1. The education law is amended by adding a new section 6438-a to read as follows: 1.