Bill S5417 version original - Jessica Scarcella-Spanton
BILL NUMBER: S5417 SPONSOR: SCARCELLA-SPANTON TITLE OF BILL: An act to amend the domestic relations law and the family court act, in relation to the effect of military service on child care PURPOSE: To allow parents deployed in the military to submit plans or requests for childcare arrangements for consideration by the court. SUMMARY OF PROVISIONS: Section 1.