Bill S5429 version original - Mario Mattera
BILL NUMBER: S5429 SPONSOR: MATTERA TITLE OF BILL: An act to amend the penal law, in relation to establishing the crimes of criminal trespass on a school bus in the first and second degrees SUMMARY OF PROVISIONS: Section 1 amends the Penal Law by adding a new section 140.18 establish- ing the crime of criminal trepass on a school bus in the second degree (class A misdeanor). Section 2 amends the Penal Law by adding a new section 140.