Bill S5384 version original - Pete Harckham
BILL NUMBER: S5384 SPONSOR: HARCKHAM TITLE OF BILL: An act to amend the transportation law, in relation to creating a road salt applicator training program PURPOSE: This legislation would create a voluntary training program for salt applicators to help minimize the negative impacts of chloride pollution on human health and the environment. SUMMARY OF PROVISIONS: Section 1 of the bill amends the transportation law by adding a new section 24 to: 1.