Bill S5397 version original - Pete Harckham
BILL NUMBER: S5397 SPONSOR: HARCKHAM TITLE OF BILL: An act to amend the vehicle and traffic law, in relation to the display of green lights on the vehicles of members of mobile crisis teams PURPOSE: Relates to the display of green lights on the vehicles of mobile crisis response staff members. SUMMARY OF PROVISIONS: This bill amends the vehicle and traffic law by allowing mobile crisis response staff members to use a green light on their motor vehicle when engaging in an emergency operation.