Bill S5371 version original - Pete Harckham
BILL NUMBER: S5371 SPONSOR: HARCKHAM TITLE OF BILL: An act to amend the mental hygiene law, in relation to funding for services of the office of addiction services and supports PURPOSE: Updates the Office of Addiction Services and Supports payment methodol- ogy to the payer for service model. SUMMARY OF PROVISIONS: Section 1 & 2 amend section 25.03 of the mental hygiene law and section 25.