Bill S5437 version original - Pete Harckham
BILL NUMBER: S5437 SPONSOR: HARCKHAM TITLE OF BILL: An act to amend the financial services law, in relation to requiring certain corporations to annually prepare a climate-related financial risk report for submission to the secretary of state and to make such report available to the public PURPOSE: To require financial institutions to prepare and publish climate risk disclosure reports. SUMMARY OF PROVISIONS: Section 1 states the legislative intent.