Bill S5484 version original - Dan Stec
BILL NUMBER: S5484 SPONSOR: STEC TITLE OF BILL: An act to amend the general construction law, in relation to the defi- nition of legislation PURPOSE: To ensure that all legislation is proposed in language that is clear and concise. SUMMARY: Section 1 amends the general construction law to require that all proposed legislation be prepared in a manner that is clear and concise. Section 2 provides the effective date.