Bill S5465A version original - Robert Jackson
BILL NUMBER: S5465 SPONSOR: JACKSON TITLE OF BILL: An act to amend the executive law, in relation to prohibiting employers from discriminating against individuals based on such individual's status as a caregiver PURPOSE: The purpose of the bill is to provide protections for caregivers under the Human Rights Law. SUMMARY OF PROVISIONS: Section 1 adds a new subdivision 42 to Section 292 of the Executive Law to define the terms "caregiver" and "care recipient.