Bill S5503 version original - Jessica Scarcella-Spanton
BILL NUMBER: S5503 SPONSOR: SCARCELLA-SPANTON TITLE OF BILL: An act to amend the public service law, in relation to creating the state office of the utility consumer advocate PURPOSE OR GENERAL IDEA OF BILL: This bill would create the State Office of the Utility Consumer Advocate (UCA). SUMMARY OF PROVISIONS: Section 1 amends the public service law by adding a new article 1-A that would establish the state Office of the Utility Consumer Advocate.