Bill S5508 version original - Jessica Scarcella-Spanton
BILL NUMBER: S5508 SPONSOR: SCARCELLA-SPANTON TITLE OF BILL: An act to amend the mental hygiene law, in relation to clarifying the standards for involuntary in-patient care and treatment PURPOSE: To provide an alternative means for mental health intervention by providing for court ordered in patient medical treatment. SUMMARY OF PROVISIONS: § 1 - adds a short title to be known as the New York State Mental Health Reform Act.