Bill S5505 version original - Jessica Scarcella-Spanton
BILL NUMBER: S5505 SPONSOR: SCARCELLA-SPANTON TITLE OF BILL: An act to amend the penal law, in relation to redefining the term "following" for a crime of stalking in the fourth degree PURPOSE: The purpose of this bill is to further define "following" under sub-sec- tion 2. SUMMARY OF PROVISIONS: Section 1: Adds a clearer definition to the term "following" pursuant to sub-division 2, to section 120.