Bill S5479 version original - Jessica Scarcella-Spanton
BILL NUMBER: S5479 SPONSOR: SCARCELLA-SPANTON TITLE OF BILL: An act to amend the penal law, in relation to elevating an assault of a retail worker to the class D felony of assault in the second degree PURPOSE: With the spike in crime throughout the country, many existing and new laws protect first responders from assault. This legislation would add a retail worker, who during the height of the COVID-19 pandemic was deemed essential, the same as a first responder.