Bill S5449 version original - Robert Jackson
BILL NUMBER: S5449 SPONSOR: JACKSON TITLE OF BILL: An act to amend the education law, in relation to requiring public schools to start no earlier than eight thirty a.m. PURPOSE: The purpose of this bill is to prohibit the start time of public schools before 8:30AM across the State.