Bill S5440 version original - Jessica Scarcella-Spanton
BILL NUMBER: S5440 SPONSOR: SCARCELLA-SPANTON TITLE OF BILL: An act to amend the general business law, in relation to the licensing of nail salons PURPOSE OF BILL: The purpose of this bill is to raise standards and improve working conditions in the nail salon industry. SUMMARY OF SPECIFIC PROVISIONS: Section 1- amends section 400 of the General Business Law to define "nail salon" as an appearance enhancement business, which practices nail specialty.