Bill S5451 version original - Robert Jackson
BILL NUMBER: S5451 SPONSOR: JACKSON TITLE OF BILL: An act to amend the education law, in relation to updating the census numbers used to calculate certain education funding PURPOSE: This legislation will update the census numbers used to calculate certain education funding. SUMMARY OF PROVISIONS: Section 1 is the title of this legislation.