Bill S5441 version original - Jessica Scarcella-Spanton
BILL NUMBER: S5441 SPONSOR: SCARCELLA-SPANTON TITLE OF BILL: An act to amend the military law, in relation to employment protections during the performance of state active duty by members of the national guard PURPOSE: The purpose of this bill is to provide employment protections during the performance of state active duty by members of the National Guard. SUMMARY OF PROVISIONS: Section 1 amends subdivision 1 of section 301 of the military law to change the definition of "military service".