Bill S5459 version original - Robert Jackson
BILL NUMBER: S5459 SPONSOR: JACKSON TITLE OF BILL: An act to amend the labor law, in relation to enacting the safeguarding employees and accountability for termination act PURPOSE: The purpose of this legislation is to require employers to have good cause when terminating an employee. SUMMARY OF PROVISIONS: Section 1 adds a new Article 20-D to the Labor Law entitled the "Safe- guarding Employees and Accountability for Termination Act" or "The SEAT Act.